requirements.txt
langchain
pypdf
python-docx
pytesseract
Pillow
PyYAML
click
tqdm
spacy
transformers
torch
accelerate
bitsandbytes